chore: Fix the build and tests
