--- packages/v2/tsconfig.build.esm.json
@@ -1,7 +1,0 @@
-{
-    "extends": "./tsconfig.build.json",
-    "compilerOptions": {
-        "module": "es6",
-        "outDir": "build/esm"
-    }
-}
