+++ packages/v2/tsconfig.json
@@ -10,8 +10,8 @@
         "esModuleInterop": true,
         "isolatedModules": true,
         "jsx": "react-jsx",
-        "lib": ["es5", "es6", "es7", "es2017", "es2018", "dom"],
-        "module": "NodeNext",
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "module": "esnext",
         "moduleResolution": "node",
         "noFallthroughCasesInSwitch": true,
         "noImplicitAny": true,
